Performance issues - FastDirectFieldAccessor
For FIELD access the relevant field is configured with field's
PropertyMetadata. That quite speeds the fields access.

A comaparison test for speed shows that we're faster than RI in
marshalling into XMLEventWriter, but slower when marshalling into
OutputStream - the reason is the internal use of XMLEvents - we should
soon do something about it.

And last thing - RI and Swe throws StackOverflow while marshalling +1000
nested attributes - maybe we should "flatten" the XML patterns?